# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{username}/{slug}/domains:
    get:
      operationId: listProjectDomains
      tags: [projects, domains]
      summary: List custom domains attached to a project
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectDomain"
    post:
      operationId: addProjectDomain
      tags: [projects, domains]
      summary: Attach a custom domain
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDomainBody"
      responses:
        "200":
          description: Created domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDomain"
  /projects/{username}/{slug}/domains/{host}:
    delete:
      operationId: removeProjectDomain
      tags: [projects, domains]
      summary: Remove a custom domain
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: host
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
  /projects/{username}/{slug}/domains/{host}/verify:
    post:
      operationId: verifyProjectDomain
      tags: [projects, domains]
      summary: Re-check verification + DNS state for a domain
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: host
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Refreshed domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDomain"
  /projects/{username}/{slug}/domains/{host}/primary:
    post:
      operationId: setPrimaryProjectDomain
      tags: [projects, domains]
      summary: Promote a verified domain to primary
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: host
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetPrimaryDomainResponse"
...
components:
  schemas:
...
    DomainVerificationRecord:
      type: object
      properties:
        type: { type: string }
        domain: { type: string }
        value: { type: string }
        reason: { type: string }
      required:
        - type
        - domain
        - value
        - reason
    DnsRecordType:
      type: string
      enum: [CNAME, A]
    DomainSuggestedRecord:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/DnsRecordType"
        name: { type: string }
        value: { type: string }
      required:
        - type
        - name
        - value
    DomainDnsMismatch:
      type: object
      properties:
        recordType:
          $ref: "#/components/schemas/DnsRecordType"
        expected: { type: string }
        actual:
          type: array
          items: { type: string }
      required:
        - recordType
        - expected
        - actual
    ProjectDomain:
      type: object
      properties:
        id: { type: string }
        host: { type: string }
        verified: { type: boolean }
        isPrimary: { type: boolean }
        misconfigured: { type: boolean }
        verificationRecords:
          type: array
          items:
            $ref: "#/components/schemas/DomainVerificationRecord"
        suggestedRecords:
          type: array
          items:
            $ref: "#/components/schemas/DomainSuggestedRecord"
        aValues:
          type: array
          nullable: true
          items: { type: string }
        cnames:
          type: array
          nullable: true
          items: { type: string }
        configuredBy:
          type: string
          nullable: true
        dnsMismatch:
          oneOf:
            - $ref: "#/components/schemas/DomainDnsMismatch"
            - type: "null"
        createdAt:
          type: string
          format: date-time
        lastCheckedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - host
        - verified
        - isPrimary
        - misconfigured
        - verificationRecords
        - suggestedRecords
        - aValues
        - cnames
        - configuredBy
        - dnsMismatch
        - createdAt
        - lastCheckedAt
    AddDomainBody:
      type: object
      properties:
        host: { type: string }
      required:
        - host
    SetPrimaryDomainResponse:
      type: object
      properties:
        status: { type: string }
      required:
        - status
...