# lib/api-zod/src/generated/types/projectDomain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DomainDnsMismatch } from "./domainDnsMismatch";
import type { DomainSuggestedRecord } from "./domainSuggestedRecord";
import type { DomainVerificationRecord } from "./domainVerificationRecord";

export interface ProjectDomain {
  id: string;
  host: string;
  verified: boolean;
  isPrimary: boolean;
  misconfigured: boolean;
  verificationRecords: DomainVerificationRecord[];
  suggestedRecords: DomainSuggestedRecord[];
  aValues: string[] | null;
  cnames: string[] | null;
  configuredBy: string | null;
  dnsMismatch: DomainDnsMismatch | null;
  createdAt: string;
  lastCheckedAt: string | null;
}